+++ lib/sequelizer/options.rb
@@ -27,7 +27,7 @@ module Sequelizer
     # the string is returned without modification
     def fix_options(passed_options)
       return passed_options unless passed_options.nil? || passed_options.is_a?(Hash)
-      sequelizer_options = db_config.merge((passed_options || {}).to_hash)
+      sequelizer_options = db_config.merge(OptionsHash.new(passed_options || {}).to_hash)
 
       if sequelizer_options[:adapter] =~ /^postgres/
         sequelizer_options[:adapter] = 'postgres'
@@ -50,12 +50,14 @@ module Sequelizer
     end
 
     # Grabs the database options from
+    #  - ~/.config/sequelizer.yml if it exists
     #  - config/database.yml if it exists
     #  - environment variables (also reads from .env)
     def db_config
       @db_config ||= begin
-        opts = OptionsHash.new(YamlConfig.new.options || {})
-        opts.merge!(EnvConfig.new.options || {})
+        opts = OptionsHash.new(YamlConfig.user_config.options)
+        opts.merge!(YamlConfig.local_config.options)
+        opts.merge!(EnvConfig.new.options)
         opts
       end
     end