+++ lib/sequelizer/yaml_config.rb
@@ -1,12 +1,33 @@
 require 'psych'
+require 'pathname'
 
 module Sequelizer
   class YamlConfig
+    attr_reader :config_file_path
+
+    class << self
+      def local_config
+        new
+      end
+
+      def user_config
+        new(user_config_path)
+      end
+
+      def user_config_path
+        Pathname.new("~") + ".config" + "sequelizer" + "database.yml"
+      end
+    end
+
+    def initialize(config_file_path = nil)
+      @config_file_path = Pathname.new(config_file_path || Pathname.pwd + "config" + "database.yml").expand_path
+    end
+
     # Returns a set of options pulled from config/database.yml
     # or +nil+ if config/database.yml doesn't exist
     def options
-      return nil unless config_file.exist?
-      config['adapter'] ? config : config[environment]
+      return {} unless config_file_path.exist?
+      config['adapter'] || config[:adapter] ? config : config[environment]
     end
 
     # The environment to load from database.yml
@@ -24,21 +45,9 @@ module Sequelizer
 
     private
 
-    # The Pathname to config/database.yml
-    def config_file
-      @config_file ||= begin
-        root + 'config' + 'database.yml'
-      end
-    end
-
-    # The root directory to search for config/database.yml
-    def root
-      @root ||= Pathname.pwd
-    end
-
     # The config as read from config/database.yml
     def config
-      @config ||= Psych.load_file(config_file)
+      @config ||= Psych.load_file(config_file_path)
     end
   end
 end