+++ test/sequelizer/test_connection_maker.rb
@@ -15,11 +15,14 @@ class TestConnectionMaker < Minitest::Test
   end
 
   def test_accepts_options_as_params
-    assert_equal :connection, Sequelizer::ConnectionMaker.new(@options).connection
+    Sequelizer::YamlConfig.stub :user_config_path, Pathname.new('/completely/made/up/path/that/does/not/exist') do
+      assert_equal :connection, Sequelizer::ConnectionMaker.new(@options).connection
+    end
   end
 
   def test_reads_options_from_yaml_config
     yaml_config = Minitest::Mock.new
+    yaml_config.expect :options, @options
     yaml_config.expect :options, @options
 
     Sequelizer::YamlConfig.stub :new, yaml_config do
@@ -31,7 +34,8 @@ class TestConnectionMaker < Minitest::Test
 
   def test_reads_options_from_env_config_if_no_yaml_config
     yaml_config = Minitest::Mock.new
-    yaml_config.expect :options, nil
+    yaml_config.expect :options, {}
+    yaml_config.expect :options, {}
 
     env_config = Minitest::Mock.new
     env_config.expect :options, @options