+++ test/sequelizer/test_options.rb
@@ -41,8 +41,15 @@ class TestOptions < Minitest::Test
   end
 
   def test_returns_a_hash_even_if_given_nil
-    options = Sequelizer::Options.new
-    assert_equal({}, options.to_hash)
+    Sequelizer::YamlConfig.stub :user_config_path, Pathname.new('/completely/made/up/path/that/does/not/exist') do
+      options = Sequelizer::Options.new
+      assert_equal({}, options.to_hash)
+    end
+  end
+
+  def test_handles_symbolized_search_path
+    options = Sequelizer::Options.new(search_path: 'passed', adapter: 'postgres')
+    assert_equal 'passed', options.search_path
   end
 end